security and access control

+++ src/Controller/MicroFilmController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\MicroFilm;
 use App\Form\MicroFilmType;
+use App\Entity\Rechercher;
 use App\Repository\MicroFilmRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,21 +16,49 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class MicroFilmController extends AbstractController
 {
+
     /**
      * @Route("/", name="micro_film_index", methods={"GET"})
      */
-    public function index(MicroFilmRepository $microFilmRepository): Response
+    public function index(MicroFilmRepository $microFilmRepository, Request $request): Response
     {
+
+        $search = new Rechercher();
+        $form = $this->createForm(SearchType::class, $search);
+        $form->handleRequest($request);
+
+
+        if ($form->isSubmitted() && $form->isValid()) {
+
+            $title = $form->getData()->getTitre();
+            $cote = $form->getData()->getCote();
+
+
+
+            if (empty($title) && empty($cote)) {
+                $this->addFlash('erreur', 'Aucun article contenant ce mot clé dans le titre n\'a été trouvé, essayez en un autre.');
+            }
+
+            return $this->render('micro_film/index.html.twig', [
+                'micro_films' => $microFilmRepository->findMicroFilm($title, $cote),
+                'form' => $form->createView()
+            ]);
+        }
+
+
         return $this->render('micro_film/index.html.twig', [
             'micro_films' => $microFilmRepository->findAll(),
+            'form' => $form->createView()
+
         ]);
     }
-
-    /**
-     * @Route("/new", name="micro_film_new", methods={"GET","POST"})
+ /**
+     * @Route("/new", name="c_d_rom_new", methods={"GET","POST"})
      */
+
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_VOLONTEER');
         $microFilm = new MicroFilm();
         $form = $this->createForm(MicroFilmType::class, $microFilm);
         $form->handleRequest($request);
@@ -63,6 +92,7 @@ class MicroFilmController extends AbstractController
      */
     public function edit(Request $request, MicroFilm $microFilm): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_VOLONTEER');
         $form = $this->createForm(MicroFilmType::class, $microFilm);
         $form->handleRequest($request);
 
@@ -83,7 +113,8 @@ class MicroFilmController extends AbstractController
      */
     public function delete(Request $request, MicroFilm $microFilm): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$microFilm->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        if ($this->isCsrfTokenValid('delete' . $microFilm->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($microFilm);
             $entityManager->flush();
